Initial commit
The current state of the code is that it works for arbitrary digraphs
implemented using NetworkX.

What is missing:

- Code to check whether or not two graphs have the same canonicalization.
This requires implementing some means of representing automorphisms. I
started implementing this using the SciPy permutation class as representatives
but this is not finished.

- Code which extracts a labelled digraph for the GRN of an input DSGRN network.

Once these are done the test cases written by Bree should run and correctness of
the implementation can be checked.